pyproject.toml
[tool.poetry]
name = "opentom-dspy"
version = "0.1.0"
description = ""
authors = ["vintro <vince@plasticlabs.ai>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
jupyterlab = "^4.1.2"
python-dotenv = "^1.0.1"
dspy-ai = "^2.1.10"
neptune = "^1.9.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
